requirements.txt
streamlit
python-docx
langchain
PyPDF2
mysql-connector-python
transformers